requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
yfinance>=0.2.36
pyyaml>=6.0
schedule>=1.2.0
pytest>=7.0.0
pytz>=2023.3
